Fix detekt findings in the Compose comment detail
Moves CommentDetailsActions to its own file, rewrites findMentionToken
without loop jumps, and drops an unused import.
